my_safe_process: try to kill the process softly first
first SIGTERM and if the process didn't die in 10 seconds, SIGKILL it.

This allows various tools like `rr`, `gcov`, `gprof`, etc to flush
their data to disk properly

+++ mysql-test/lib/My/SafeProcess/safe_process.cc
@@ -144,7 +144,12 @@ static int kill_child(bool was_killed)
   message("Killing child: %d", child_pid);
   // Terminate whole process group
   if (! was_killed)
-    kill(-child_pid, SIGKILL);
+  {
+    kill(-child_pid, SIGTERM);
+    sleep(10); // will be interrupted by SIGCHLD
+    if (!waitpid(child_pid, &status, WNOHANG))
+      kill(-child_pid, SIGKILL);
+  }
 
   pid_t ret_pid= waitpid(child_pid, &status, 0);
   if (ret_pid == child_pid)
